Added flac support to player. DataURI is now generated on the fly.
For some strange reason Chrome does not support playing files with
mimetype audio/x-flac but supports audio/flac which are essentially the
same so added a function to set the proper mimetype. Also DB was getting
too large by storing DataURI in it so it is now generated on the fly.